blog-project/blogsite/requirements.txt
alembic==0.9.3.dev0
amqp==1.4.9
anyjson==0.3.3
apturl==0.5.2
asn1crypto==0.24.0
astroid==2.3.3
astropy==3.1.2
attrs==19.3.0
autopep8==1.4.4
awscli==1.16.191
Babel==2.4.0
backcall==0.1.0
billiard==3.3.0.23
bleach==3.1.4
blinker==1.4
boto3==1.9.181
botocore==1.12.181
Brlapi==0.6.6
celery==3.1.18
certifi==2019.9.11
chardet==3.0.4
click==6.7
colorama==0.3.9
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distro-info===0.18ubuntu0.18.04.1
Django==2.0
django-celery-results==1.1.2
django-cors-headers==2.4.0
django-filter==2.1.0
django-stubs==1.5.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
docopt==0.4.0
docutils==0.15.2
elasticache-auto-discovery==1.0.0
entrypoints==0.3
flake8==3.7.9
Flask==0.12.2
Flask-BabelEx==0.9.3
Flask-Compress==1.4.0
Flask-Gravatar==0.4.2
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.1.1
Flask-Paranoid==0.2.0
Flask-Principal==0.4.0
Flask-Security==1.7.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.14.2
httpie==0.9.8
httplib2==0.9.2
idna==2.8
importlib-metadata==1.6.0
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==0.24
jdcal==1.4
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.4
jplephem==2.12
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
keyring==10.6.0
keyrings.alt==3.0
kombu==3.0.37
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
louis==3.5.0
Mako==1.0.7
mandrill==1.0.57
MarkupSafe==1.1.1
mate-hud==18.4.8
mate-tweak==18.4.16
mccabe==0.6.1
meld==3.18.0
mistune==0.8.4
more-itertools==7.0.0
mypy==0.770
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.0.5
netifaces==0.10.4
nose==1.3.7
notebook==6.0.3
numpy==1.16.2
oauth==1.0.1
olefile==0.45.1
onboard==1.4.1
pandocfilters==1.4.2
paramiko==2.0.0
parso==0.6.2
passlib==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==5.1.0
pkginfo==1.5.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
psutil==5.4.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
ptyprocess==0.6.0
pyasn1==0.4.5
pycairo==1.16.2
pycodestyle==2.5.0
pycrypto==2.6.1
pycups==1.9.73
pyflakes==2.1.1
Pygments==2.6.1
pygobject==3.26.1
pyinotify==0.9.6
PyJWT==1.6.4
pylint==2.4.4
pymemcache==2.2.2
pymongo==3.9.0
pyOpenSSL==17.5.0
pyrsistent==0.16.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-xapp==1.0.0
python-xlib==0.20
pytz==2018.5
pyxattr==0.6.0
pyxdg==0.25
PyYAML==5.1
pyzmq==19.0.0
readme-renderer==24.0
redis==3.4.1
reportlab==3.4.0
requests==2.22.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
rsa==3.4.2
s3transfer==0.2.1
SecretStorage==2.3.1
Send2Trash==1.5.0
setproctitle==1.1.10
sgp4==1.4
Shapely==1.6.4.post2
simplejson==3.13.2
six==1.14.0
skyfield==1.16
SQLAlchemy==1.1.11
sqlparse==0.3.0
sshtunnel==0.1.4
systemd-python==234
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.38.0
traitlets==4.3.3
twine==2.0.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzwhere==3.0.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.7
usb-creator==0.3.3
uWSGI==2.0.18
vine==1.3.0
virtualenv==16.7.7
wadllib==1.3.2
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.1
xkit==0.0.0
youtube-dl==2018.3.14
zipp==3.1.0
zope.interface==4.3.2
